Group similar functionality in parse code
This makes the code easier to read. As well as limits the amount of '+1', '-1', etc. actions.
We don't create new name, sequence and qualities variables until it is certain that
the FASTQ record has the complete 4 newlines.
This code is slightly faster (5-10% in my benchmarks). Maybe this has also to do with
the branch predictor in the CPU having an easier time because similar functionality is grouped.