Infer qualities end from sequence length
For each character in the input, two checks were performed. Is it a newline, and have we reached bufend yet?
With this change the lenght of the quality string is inferred from the sequence string. Therefore we can skip
sequence length x 2 checks. We just perform checks afterward to see if we end on a newline.

This has a slight risk, as there are some possible outcomes:
1. The newline is indeed the last newline of the record.
2. The newline is a newline in a following record, because the quality string was truncated and we were unlucky.

In case of 1. Success, but in case of 2. there are two possible outcomes:
1. It is the ending newline of the name, sequence or second header line.
   In which case the parser will throw an error for the next record.
   Which is good, problem averted.
2. It is the ending newline of the next record's quality string. In that case
   we have included an entire record in this record's quality string and the
   parser keeps happily parsing.

Again in case of 2. This leads to a missing record, and a record with
invalid quality values. This might be catched by upstream tools
(problem averted!), or it will lead to wrong quality values for this record.
Probably one misrepresented record will not lead to a different analysis outcome.