Avoid relative seek
The value returned by file.tell() is an opaque number (and is
documented as such in the Python docs). On Windows, the
"if file.tell() > 0" was unreliable because file.tell() was
actually negative.

Instead, just save the original position with file.tell(), then
return to it after reading the byte we are interested in.