Run tests in the project directory
The default recommended by cibuildwheel is to not run the tests from the
project directory in order to ensure that the installed wheel is tested, not
the code in the repo, but we already avoid this problem by using the src/
layout.

The tests currently assume they are run from the root of the repo. Changing
this would be another option, but requires more intrusive changes.