Fix unoptimized wheels
setuptools starting from version 75.7.0 changed how CFLAGS work: They no
longer are added to the CFLAGS Python was compiled with, but replace them.

With this change, our CFLAGS of `-g0 -DNDEBUG` get used directly and the
wheels are compiled without proper optimizations because `-O2` is missing.

Within cibuildwheel, the Python compiler options are:

    -fno-strict-overflow -Wsign-compare -DNDEBUG -g -O3 -Wall

Of these, we really only require `-O3` and `-DNDEBUG`, so the easiest fix
should be to just add `-O3` to the CFLAGS that we provide.

Found by @rhpvorderman in https://github.com/marcelm/cutadapt/issues/843